+++ .storybook/components/DesignTokens/Tier2/TypographyUsage.js
@@ -6,15 +6,18 @@ import { PRESET_SIZE_MAP } from '../Tier1/TypographyPresets';
 
 export class Tier2TypographyUsage extends Component {
   render() {
-    const renderTypeToken = (size, preset) => {
+    const renderTypeToken = (size, preset, legacySize) => {
       // remove variant suffix
       const normalizedPreset = preset.split('-')[0];
       const { fontSize, lineHeight } = PRESET_SIZE_MAP[normalizedPreset];
 
+      const name =
+        `eds-theme-typography-${size}` + (legacySize ? `/ ${legacySize}` : '');
+
       return (
         <TokenSpecimen
           comment={`Ref: preset-${preset}`}
-          name={`eds-theme-typography-${size}`}
+          name={name}
           specimenClassName={`u-theme-typography-${size}`}
           value={`${fontSize} / ${lineHeight}`}
           variant="typography-body"
@@ -26,16 +29,16 @@ export class Tier2TypographyUsage extends Component {
       <div>
         <Section title="Headline">
           <Grid>
-            {renderTypeToken('headline-lg', '001-bold')}
-            {renderTypeToken('headline-md', '002-bold')}
-            {renderTypeToken('headline-sm', '003-bold')}
+            {renderTypeToken('headline-lg', '001-bold', 'h1')}
+            {renderTypeToken('headline-md', '002-bold', 'h2')}
+            {renderTypeToken('headline-sm', '003-bold', 'h3')}
           </Grid>
         </Section>
 
         <Section title="Title">
           <Grid>
-            {renderTypeToken('title-md', '004-bold')}
-            {renderTypeToken('title-sm', '005-bold')}
+            {renderTypeToken('title-md', '004-bold', 'h4')}
+            {renderTypeToken('title-sm', '005-bold', 'h5')}
           </Grid>
         </Section>