+++ src/components/HorizontalStepper/HorizontalStepper.tsx
@@ -20,8 +20,6 @@ export type Props = {
 };
 
 /**
- * BETA: This component is still a work in progress and is subject to change.
- *
  * ```ts
  * import {HorizontalStepper} from "@chanzuckerberg/eds";
  * ```